production/backend/requirements.txt
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework_simplejwt==5.0.0
django-cors-headers==3.10.0
Markdown==3.3.4
